% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainCoxRel}
\alias{mainCoxRel}
\title{This is the main study dataset (mainCoxRel)}
\format{
A data frame with 1500 rows and 5 variables:
\describe{
  \item{x}{mismeasured variable x , continuous}
  \item{s1}{confounder, continuous, perfectly measured }
  \item{s2}{confounder, continuous,perfectly measured }
  \item{Y}{time, continuous}
  \item{failed}{event, binary}
}
}
\usage{
mainCoxRel
}
\description{
This is the main study dataset (mainCoxRel)
}
\examples{
data("mainCoxRel", package = "merc")
}
\keyword{datasets}
