% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainLinearRel}
\alias{mainLinearRel}
\title{This is the main study dataset (mainLinearRel)}
\format{
A data frame with 1500 rows and 5 variables:
\describe{
  \item{x}{mismeasured variable x , continuous}
  \item{z}{mismeasured variable z , continuous}
  \item{s1}{confounder, continuous, perfectly measured }
  \item{s2}{confounder, binary, perfectly measured }
  \item{Y}{outcome, continuous}
}
}
\usage{
mainLinearRel
}
\description{
This is the main study dataset (mainLinearRel)
}
\examples{
data("mainLinearRel", package = "merc")
}
\keyword{datasets}
