% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainLogRel}
\alias{mainLogRel}
\title{This is the main study dataset (mainLogRel) of logistic regression}
\format{
A data frame with 1500 rows and 3 variables:
\describe{
  \item{x}{mismeasured variable, continuous}
  \item{s}{confounder, binary, perfectly measured }
  \item{Y}{outcome, continuous}
}
}
\usage{
mainLogRel
}
\description{
This is the main study dataset (mainLogRel) of logistic regression
}
\examples{
data("mainLogRel", package = "merc")
}
\keyword{datasets}
