% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainRelib1}
\alias{mainRelib1}
\title{This is the main study dataset (mainRelib1)}
\format{
A data frame with 5000 rows and 14 variables:
\describe{
  \item{x}{mismeasured variable, continuous}
  \item{s}{confounder, binary, perfectly measured }
  \item{Y}{outcome, continuous}
}
}
\usage{
mainRelib1
}
\description{
This is the main study dataset (mainRelib1)
}
\examples{
data("mainRelib1", package = "merc")
}
\keyword{datasets}
