% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainRelib2}
\alias{mainRelib2}
\title{This is the main study dataset (mainRelib1)}
\format{
A data frame with 5000 rows and 14 variables:
\describe{
  \item{x}{mismeasured variable x , continuous}
  \item{z}{mismeasured variable z , continuous}
  \item{s1}{confounder, continuous, perfectly measured }
  \item{s2}{confounder, binary, perfectly measured }
  \item{Y}{outcome, continuous}
}
}
\usage{
mainRelib2
}
\description{
This is the main study dataset (mainRelib1)
}
\examples{
data("mainRelib2", package = "merc")
}
\keyword{datasets}
