% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relib1}
\alias{relib1}
\title{This is a simulated Reliability study.
Logistic Regression on one covariate measured with error and one confounder， 4 replicates [reliability study]}
\format{
A data frame with 1500 rows and 4 reliability measures for the mismeasured variable x:
\describe{
  \item{x1}{first reliability measure for the covariate measured with error}
  \item{x2}{second reliability measure for the covariate measured with error}
  \item{x3}{third reliability measure for the covariate measured with error}
  \item{x4}{fourth reliability measure for the covariate measured with error}
  }
}
\usage{
relib1
}
\description{
This is the reliability dataset (relib1)
}
\examples{
data("relib1", package = "merc")
}
\keyword{datasets}
