% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relib2}
\alias{relib2}
\title{This is a simulated Reliability study.
Linear Regression on two covariates measured with error and two confounders , 2 replicates
This is the reliability dataset (relib2)}
\format{
A data frame with 1500 rows and 2 reliability measures for each 2 mismeasured variables
\describe{
  \item{x1}{first reliability measure for the mismeasured x}
  \item{z1}{first reliability measure for the mismeasured z}
  \item{x2}{second reliability measure for the mismeasured x}
  \item{z2}{fourth reliability measure for the mismeasured z}
  }
}
\usage{
relib2
}
\description{
This is a simulated Reliability study.
Linear Regression on two covariates measured with error and two confounders , 2 replicates
This is the reliability dataset (relib2)
}
\examples{
data("relib2", package = "merc")
}
\keyword{datasets}
