% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relibCox}
\alias{relibCox}
\title{This is a simulated Reliability study.
Cox Regression on one covariate measured with error and two confounders , 4 replicates
This is the reliability dataset (relibCox)}
\format{
A data frame with 750 rows and 4 reliability measures for one mismeasured variable
\describe{
  \item{x1}{first reliability measure for the covariate measured with error}
  \item{x2}{second reliability measure for the covariate measured with error}
  \item{x3}{third reliability measure for the covariate measured with error}
  \item{x4}{fourth reliability measure for the covariate measured with error}
  }
}
\usage{
relibCox
}
\description{
This is a simulated Reliability study.
Cox Regression on one covariate measured with error and two confounders , 4 replicates
This is the reliability dataset (relibCox)
}
\examples{
data("relibCox", package = "merc")
}
\keyword{datasets}
