% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relibpls.R
\name{relibpls}
\alias{relibpls}
\title{relibpls}
\usage{
relibpls(
  supplyEstimates = FALSE,
  relib,
  pointEstimates = NA,
  vcovEstimates = NA,
  sur,
  woe = NA,
  outcome = NA,
  weri,
  rr,
  ms,
  weights,
  r1 = 1,
  method = "lm",
  family = NA,
  link = NA
)
}
\arguments{
\item{relib}{Name of the reliability dataset}

\item{pointEstimates}{A numeric vector of point estimates from uncorrected standard regression analysis.
It must be in the order of the variables indicated in sur and woe
Intercept estimate must be removed. Only binary and numeric variables are accepted, all other classes of variables must be transformed
(this is automatically done by most regression softwares). Must include names for each point estimates corresponding to the (transformed)
names from `covCalib` followed by `covOutcome`. Must be supplied if supplyEstimates = TRUE.}

\item{vcovEstimates}{A p by p Variance-covariance matrix estimates from uncorrected standard regression analysis.
Intercept estimates must be removed. Only binary and numeric variables are accepted, all other classes of variables must be transformed
(this is automatically done by most regression softwares). Must include column names (excluding intercept) for the estimates corresponding
to the (transformed) names from `covCalib` followed by `covOutcome`. Must be supplied if supplyEstimates = TRUE.}

\item{sur}{character vector of mismeasured exposure and covariates (i.e. surrogates) in the main study dataset}

\item{woe}{Character vector of names of perfectly measured covariates}

\item{outcome}{Outcome variable}

\item{weri}{Character vector of names of variables which have reliability measures. These should have a set of variable names for each reliability measure
(e.g. x1 y1 z1 x2 y2 z2 for measures). The variables must be in the same order as the names in the names and increments dataset specified in `pointEstimates`}

\item{rr}{Number of reliability measures taken (e.g. 2).}

\item{ms}{name of main study data set}

\item{weights}{Name of the dataset containing the variable names and increments for the odds ratio or regression slopes.Including mismeasured and perfectly measured variables}

\item{method}{Methods for modeling, currently only `lm` or `glm` methods are available. Required.}

\item{family}{Supply family parameter to pass to glm function. Not a character. Required if method="glm".}

\item{link}{Supply link parameter to pass to glm function. Should be character. Required if method="glm".}

\item{supplyEstimates=FALSE}{Indicates whether uncorrected estimates will be supplied by the user.}

\item{r1=1}{Number of replicates in main study}
}
\value{
printable dataframe from standard regression results (when supplyEstimates==FALSE) as well as corrected results
}
\description{
This funciton calculates regression coefficients, their standard errors, and odds ratios,
when relevant, and 95% confidence intervals for a biologically meaningful difference specified
by model covariates. Logistic model are implemented. A reliability study is required to empirically
charaterize the measurement error model. Details are given in Rosner et al.(1989), Rosner et al.(1990), and
Rosner et al.(1992) including "real data" examples
}
\examples{
# # Only one mismeasured covariate, linear model
y <- c(1,1,2,3,2,3)
x <- c(2,2.1,3,4,2.9,4.1)
s <- c(2,5,4,3,3,5)
case <- c(0,1,0,1,0,0)
age <- c(10,10,11,12,13,13)
test <- data.frame(y,x,s,case,age)

x <- c(1.1,1.2,0.8)
x2 <- c(1.2,1.2,0.9)
x3 <- c(0.9,1.0,1.3)
relib <- data.frame(x,x2,x3)
wts <- data.frame(x=1,s=1)

### Fit main study linear model
outModel <- lm(y ~ x + s , data = test)
outcomeParam=coef(outModel)
outcomeParamVCOV=vcov(outModel)
outcomeModelResults<-(list(outcomeParam,outcomeParamVCOV))
Bstar<-outcomeParam[2:length(outcomeParam)] #p' x 1
VBstar<-outcomeParamVCOV[2:length(outcomeParam),2:length(outcomeParam)]
relibpls(supplyEstimates=TRUE, relib=relib, pointEstimates = Bstar, vcovEstimates = VBstar, sur = c("x"), woe = c("s"), weri = c("x","x2","x3"), rr=3, ms=test, weights=wts, method = "lm" )
relibpls(supplyEstimates=FALSE, relib = relib, sur = c("x"), woe = c("s"), weri = c("x","x2","x3"), outcome = c("y"), rr=3, ms=test,method = "lm",weights=wts)

# # Only one mismeasured covariate, logistic model
### Fit main study logistic model
outModel <- glm(case ~ x + s , family = binomial(link="logit"), data = test)
outcomeParam=coef(outModel)
outcomeParamVCOV=vcov(outModel)
outcomeModelResults<-(list(outcomeParam,outcomeParamVCOV))
Bstar<-outcomeParam[2:length(outcomeParam)] #p' x 1
VBstar<-outcomeParamVCOV[2:length(outcomeParam),2:length(outcomeParam)] # p' x p'
relibpls(supplyEstimates=TRUE, relib=relib, pointEstimates = Bstar, vcovEstimates = VBstar, sur = c("x"), woe = c("s"), weri = c("x","x2","x3"), rr=3, ms=test, weights=wts,link = "logit",method = "glm" )
relibpls(supplyEstimates=FALSE, relib = relib, sur = c("x"), woe = c("s"), weri = c("x","x2","x3"), outcome = c("case"), rr=3, ms=test,method = "glm", family = binomial, link = "logit", weights=wts)
}
\references{
Rosner B, Spiegelman D, Willett WC. Correction of logistic regression relative risk estimates and confidence intervals for measurement error: the case of multiple covariates measured with error. Am J Epidemiol. 1990 Oct;132(4):734-45. doi: 10.1093/oxfordjournals.aje.a115715. PMID: 2403114.
}
\author{
Xiaofan Liu and Xin Zhou
}
